Bugfix: Framework couldn't create largest IOQ's.
In QEMU, dnvme and in tnvme all Q sizes, i.e. the number of elements in a Q,
was represented by an uint16_t. Thus the largest Q couldn't be created. A
conversion from uint16_t to uint32_t resolved this issue. However, the
recurcussions touched many files.

+++ GrpBasicInit/writeDataPat_r10b.cpp
@@ -147,9 +147,9 @@ void
 WriteDataPat_r10b::SendToIOSQ(SharedIOSQPtr iosq, SharedIOCQPtr iocq,
     SharedWritePtr writeCmd, string qualifier)
 {
-    uint16_t numCE;
-    uint16_t ceRemain;
-    uint16_t numReaped;
+    uint32_t numCE;
+    uint32_t ceRemain;
+    uint32_t numReaped;
     uint32_t isrCount;
