+++ GrpNVMReadCmd/limitedRetry_r10b.cpp
@@ -80,7 +80,7 @@ LimitedRetry_r10b::RunCoreTest()
      * 1) Test CreateResources_r10b has run prior.
      * \endverbatim
      */
-    uint16_t numCE;
+    uint32_t numCE;
     uint32_t isrCountB4;
