+++ GrpQueues/ioqRollChkSame_r10b.h
@@ -59,9 +59,9 @@ private:
     ///////////////////////////////////////////////////////////////////////////
     // Adding a member variable? Then edit the copy constructor and operator=().
     ///////////////////////////////////////////////////////////////////////////
-    void IOQRollChkSame(uint16_t numEntriesIOQ);
+    void IOQRollChkSame(uint32_t numEntriesIOQ);
     SharedWritePtr SetWriteCmd();
-    void ReapAndVerifyCE(SharedIOCQPtr iocq, uint16_t expectedVal);
+    void ReapAndVerifyCE(SharedIOCQPtr iocq, uint32_t expectedVal);
     void VerifyQPointers(SharedIOSQPtr iosq, SharedIOCQPtr iocq);
 };
